requirements.txt
# Core
pydantic>=2.0
pydantic-settings>=2.0

# LLM Providers
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
openpyxl>=3.1.0
